+++ Tools/Math.h
@@ -1,15 +1,15 @@
 #pragma once
 
 // The ratio of the circumference of a circle to its diameter.
-constexpr float PI = 3.14159265358979323846l;
+constexpr double PI = 3.14159265358979323846l;
 
 // Converts an angle measured in degrees to an approximately equivalent angle measured in radians.
-inline constexpr float toRadians(const float angdeg) {
+inline constexpr double toRadians(const double angdeg) {
   return PI / 180 * angdeg;
 }
 
 // Converts an angle measured in radians to an approximately equivalent angle measured in degrees.
-inline constexpr float toDegrees(const float angrad) {
+inline constexpr double toDegrees(const double angrad) {
   return 180 / PI * angrad;
 }
